added buyer accepts marketing to wpm
