+++ packages/web-pixels-extension/src/schemas/pixel-events.ts
@@ -531,6 +531,13 @@ export const pixelEvents = {
               'The billing address to where the order will be charged.',
           },
         },
+        buyerAcceptsEmailMarketing: {
+          type: 'boolean',
+          metadata: {
+            description:
+              "Returns true if the customer checks the email marketing subscription checkbox. Returns false if not. This property is only available if you've [upgraded to Checkout Extensibility](https://help.shopify.com/en/manual/checkout-settings/checkout-extensibility/checkout-upgrade).",
+          },
+        },
         currencyCode: {
           type: 'string',
           metadata: {
