+++ packages/web-pixels-extension/src/types/ExtensionApi.ts
@@ -1,11 +1,11 @@
 import type {Browser} from './PixelEvents';
-import type {EventBus} from './EventBus';
+import type {WebPixelsManager} from './WebPixelsManager';
 import type {RegisterInit} from './RegisterInit';
 
 export interface ExtensionApi {
   readonly settings: Record<string, any>;
   readonly analytics: {
-    readonly subscribe: EventBus['subscribe'];
+    readonly subscribe: WebPixelsManager['subscribe'];
   };
   readonly browser: Browser;
   readonly init: RegisterInit;
