+++ packages/web-pixels-extension/src/types/PixelEvents/index.ts
@@ -779,6 +779,15 @@ export interface Checkout {
   billingAddress: MailingAddress | null;
 
   /**
+   * Returns true if the customer checks the email marketing
+   * subscription checkbox. Returns false if not. This
+   * property is only available if you've [upgraded to Checkout
+   * Extensibility](https://help.shopify.com/en/manual/checkout-
+   * settings/checkout-extensibility/checkout-upgrade).
+   */
+  buyerAcceptsEmailMarketing: boolean;
+
+  /**
    * The three-letter code that represents the currency, for example, USD.
    * Supported codes include standard ISO 4217 codes, legacy codes, and non-
    * standard codes.
