+++ packages/web-pixels-extension/src/types/PublicApi.ts
@@ -1,8 +1,8 @@
-import type {EventBus} from './EventBus';
+import type {WebPixelsManager} from './WebPixelsManager';
 import type {VisitorApi} from './VisitorApi';
 
 export interface PublicApi {
-  publish: EventBus['publish'];
-  publishCustomEvent: EventBus['publishCustomEvent'];
+  publish: WebPixelsManager['publish'];
+  publishCustomEvent: WebPixelsManager['publishCustomEvent'];
   visitor: VisitorApi['visitor'];
 }
